schema(thread): check Thread_CreateThread permission
