+++ __tests__/schema/thread/create-thread.ts
@@ -28,6 +28,7 @@ import {
   assertSuccessfulGraphQLQuery,
   createMessageHandler,
   createTestClient,
+  createUuidHandler,
 } from '../../__utils__'
 import { mockEndpointsForThreads } from './thread'
 
@@ -59,6 +60,7 @@ test('unauthenticated user gets error', async () => {
 test('thread gets created, cache mutated as expected', async () => {
   const client = createTestClient({ userId: user.id })
 
+  // TODO: this also needs to include the user. Either as an optional argument or change up `mockEndpointsForThreads` so that it does not override other handlers...
   mockEndpointsForThreads(article, [[comment]])
 
   //fill cache
